// apps/api/src/generated/prisma/models/Session.ts
export type SessionModel = runtime.Types.Result.DefaultSelection<Prisma.$SessionPayload>
// ...
export type $SessionPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Session"
  objects: {
    user: Prisma.$UserPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    userId: string
    hashedToken: string
    expiresAt: Date
    createdAt: Date
    userAgent: string | null
    ip: string | null
  }, ExtArgs["result"]["session"]>
  composites: {}
}